Make utility function more generally available
